Feat: Separate Configs to models
