+++ crates/infrarust/src/core/config/models/access_list.rs
@@ -1,0 +1,8 @@
+use serde::Deserialize;
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct AccessListConfig<T> {
+    pub enabled: bool,
+    pub whitelist: Vec<T>,
+    pub blacklist: Vec<T>,
+}
